Rename IImage to ResponsiveImage to better fit OW4 model
